Estimate the Empirical Probability of Whether Two Set of Cells from Distinct Batches Belong to the Same Population

This function computes the empirical probability that two sets of cells from distinct batches belong to the same population, based on the output of getIDEr.

Usage

estimateProb(
  seu,
  ider,
  batch.var = "Batch",
  n_size = 40,
  n.perm = 5,
  verbose = FALSE
)

Arguments

seu

A Seurat object.

ider

A list returned by the getIDEr function.

batch.var

Character string specifying the metadata column that contains batch information. Default is "Batch".

n_size

Numeric value indicating the number of cells per group used to compute the similarity. Default is 40.

n.perm

Numeric value specifying the number of permutations to perform.

verbose

Logical. If TRUE, progress messages are printed. Default is FALSE.

Value

A Seurat object with additional columns for the IDER-based similarity and the empirical probability of rejection.

See also