Final Clustering Step for Meta-Clustering

This function merges initial clusters into final clusters based on the IDEr similarity matrix.

Usage

finalClustering(
  seu,
  dist,
  cutree.by = "h",
  cutree.h = 0.45,
  cutree.k = 3,
  hc.method = "complete"
)

Arguments

seu

A Seurat object that has undergone the getIDEr step. Required.

dist

A list output from the getIDEr function. Required.

cutree.by

Character string specifying whether to cut the dendrogram by height ("h") or by a fixed number of clusters ("k"). Default is "h".

cutree.h

Numeric value between 0 and 1 indicating the height at which to cut the dendrogram. This parameter is ignored if cutree.by = "k". Default is 0.45.

cutree.k

Numeric value specifying the number of clusters to generate if cutree.by = "k". This parameter is ignored if cutree.by = "h". Default is 3.

hc.method

Character string specifying the method to be used in hierarchical clustering (passed to hclust).

Value

A Seurat object with the final clustering results stored in the CIDER_clusters column of its meta.data.

See also