Estimate the empirical probability of whether two set of cells from distinct batches belong to the same population

Estimate the empirical probability of whether two set of cells from distinct batches belong to the same population

estimateProb(seu, ider, n_size = 40, n.perm = 5, verbose = FALSE)

Arguments

seu

A Seurat object

ider

The output list of function `getIDEr`.

n_size

Number of cells per group used to compute the similarity. Default: 40

n.perm

Numeric. Time of permutations.

verbose

Boolean. Print out progress or not. (Default: FALSEW)

Value

A Seurat object with IDER-based similarity and empirical probability of rejection

See also